feat(#777): wire VCR-VER-003 addressing validator UNCONDITIONALLY into main.rs
Move the validator from synth-verify into synth-core::static_data_addr (single
source; synth-verify::addr re-exports it) and call it in the mixed-split reloc
retargeting path (main.rs). It runs on EVERY compilation — the shipping build
is `--features riscv`, NOT `verify`, so a verify-gated check would stay dormant
in exactly the build that shipped #757 four times.

The retarget loop now records the EMITTED (seg_index k, addend) per static-data
reloc — never recomputed — and after the loop validate_reloc_resolutions()
proves each resolves to the runtime-correct byte (segments applied in
declaration order, later-wins). A wrong-segment resolution hard-errors the
compile (anyhow::bail!) with the offending symbol/addend/served/runtime bytes.

Red-first demonstration (end-to-end through main.rs), captured for the PR:
  * `.rposition()` (shipped): gale's real loom.wasm compiles clean.
  * revert to `.position()`: compile FAILS —
      func 9 reloc @ 0x250: __synth_wasm_seg_0+0x8 -> linmem 0x100008
      serves 0x02 but the runtime image owns 0x67 ('g')
    exactly gale's historical got=[2,0,0,0,..] = __synth_wasm_seg_0+8 (#757).
  * restore `.rposition()`: clean again.

Durable regression gate: crates/synth-cli/tests/vcr_ver_003_addr_777.rs compiles
the REAL gale module and asserts exit 0 — because the validator is wired
unconditionally, this GREEN goes RED the instant someone reverts the fix. Honours
the "the real module must live in CI; 7 synthetic shapes were all green" lesson.
The synthetic 3-overlap DISCRIMINATION (same validator RED on .position(), GREEN
on .rposition()) is the permanent unit test in synth-core.

Concrete byte-equality, not fabricated SMT: the truth image is reconstructed
independently of the chosen segment, so mirror-pinning vacuity is structurally
excluded. Part of VCR-VER-003 (epic #242).
